+++ testing/integration/src/lib.rs
@@ -1,11 +1,11 @@
 pub mod common;
 
 #[cfg(test)]
-pub mod integration_tests;
+pub mod consensus_integration_tests;
 
 #[cfg(test)]
-pub mod pipeline_tests;
+pub mod consensus_pipeline_tests;
 
 #[cfg(test)]
 #[cfg(feature = "devnet-prealloc")]
-pub mod daemon_benchmarks;
+pub mod mempool_benchmarks;
